refactor: split ontology controller into entities/versions/drift/admin package per ADR-0008 (#2253)
